+++ at_contacts_group_flutter/lib/desktop_routes/desktop_routes.dart
@@ -96,6 +96,7 @@ class DesktopGroupSetupRoutes {
           },
           onBackArrowTap: args['onBackArrowTap'],
           onDoneTap: args['onDoneTap'],
+          contactSelectedHistory: args['contactSelectedHistory'],
         );
       }
     };